requirements.txt
discord.py==2.4.0
pytest==8.3.3
PyYAML==6.0.2
setuptools==75.1.0

requests~=2.32.3
pytz~=2024.2
yfinance~=0.2.43
discord~=2.3.2
APScheduler~=3.10.4
spotipy~=2.24.0